requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi~=2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography~=44.0.2
Django~=5.1.6
django-allauth~=65.4.1
django-apscheduler==0.7.0
django-filter~=25.1
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
tzlocal~=5.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
pip~=24.3.1
pytz~=2025.1
django-debug-toolbar~=5.0.1
djangorestframework~=3.16.0
django-modeltranslation~=0.19.12